style: address SwiftLint --strict findings on PR #105
- static_over_final_class (×2): use `override static func` in the
  final `MockURLProtocol` subclass for `canInit` / `canonicalRequest`.
- file_length: FoodItem.swift was 455 lines after adding the detail-
  response init (the sample fixtures dominate). Extracted the new
  `init(from: FoodDetailResponse)` into FoodItem+FoodDetailResponse.swift
  as a convenience initializer delegating to the designated init.
